requirements.txt
chardet
baidu-aip
httpx
pyinstaller
PyQt5-stubs
pyqt5-tools
pywin32
retrying
system-hotkey
xcffib
xpybutil